Moved DownloaderRegistry to datasetinsights.io.source.base.py
