+++ datasetinsights/datasets/__init__.py
@@ -1,10 +1,11 @@
+from datasetinsights.io.source.unity_simulation import UnitySimulationDownloader
+
 from .base import Dataset
 from .cityscapes import Cityscapes
 from .coco import CocoDetection
 from .groceries_real import GoogleGroceriesReal, GroceriesReal
 from .nyudepth import NyuDepth
 from .synthetic import SynDetection2D, read_bounding_box_2d
-from .unity_simulation import UnitySimulationDownloader
 
 __all__ = [
     "Cityscapes",
